Pass local login & register test
Login page and register modal should be usable new
