+++ src/App.js
@@ -1,12 +1,22 @@
-import { Layout } from "antd";
+import { Layout, Menu, Dropdown, Button } from "antd";
+import { UserOutlined } from "@ant-design/icons";
 import React from "react";
-import LoginPageAlter from "./components/LoginPageAlter";
+import LoginPage from "./components/LoginPage";
 
 const { Component } = React;
 const { Header, Content } = Layout;
 class App extends Component {
+  state = {
+    authed: false,
+    asManager: false,
+  }
   componentDidMount() {
-
+    const authToken = localStorage.getItem("authToken");
+    const asManager  = localStorage.getItem("asManager") === "true";
+    this.setState({
+      authed: authToken !== null,
+      asManager, 
+    });
   }
 
   handleLoginSuccess = (token, asManager) => {
@@ -18,9 +28,25 @@ class App extends Component {
     });
   };
 
+  handleLogOut = () => {
+    localStorage.removeItem("authToken");
+    localStorage.removeItem("asManager");
+    this.setState({
+      authed: false,
+    });
+  }
+
+  userMenu = (
+    <Menu>
+      <Menu.Item key="logout" onClick={this.handleLogOut}>
+        Log Out
+      </Menu.Item>
+    </Menu>
+  );
+
   renderContent() {
-    if (true) {
-      return <LoginPageAlter />
+    if (!this.state.authed) {
+      return <LoginPage handleLoginSuccess={this.handleLoginSuccess}/>
     }
   }
   render() {
@@ -30,6 +56,13 @@ class App extends Component {
           <div style={{ fontSize: 30, fontWeight: 600, color: "white", lineHeight: "96px" }}>
             SweetHome
           </div>
+          {this.state.authed && (
+            <div>
+              <Dropdown trigger="click" overlay={this.userMenu}>
+                <Button icon={<UserOutlined />} shape="circle" />
+              </Dropdown>
+            </div>
+          )}
         </Header>
         <Content>
           {this.renderContent()}
